@@ -4,7 +4,7 @@ import com.baomidou.mybatisplus.annotation.*;
 import com.tongyi.common.utils.JsonUtils;
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 import java.io.Serializable;
 import java.util.Date;
 
@@ -18,7 +18,7 @@ public class Article implements Serializable {
     @TableId(type = IdType.AUTO)
     private int id;
     private int type;
-    @TableField(insertStrategy = FieldStrategy.IGNORED)//title重复则不插入
+    @TableField(insertStrategy = FieldStrategy.NOT_NULL)//title重复则不插入
     @NotEmpty(message = "标题不得为空")
     private String title;
     private String tags;
