@@ -2,9 +2,9 @@ package com.tongyi.modules.wx.form;
 
 import lombok.Data;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
 
 @Data
 public class WxQrCodeForm {
