@@ -22,7 +22,7 @@ import com.tongyi.modules.sys.entity.SysMenuEntity;
 import com.tongyi.modules.sys.entity.SysOrgEntity;
 import com.tongyi.modules.sys.entity.SysUserEntity;
 import com.tongyi.modules.sys.service.*;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
