@@ -11,7 +11,7 @@
  */
 package com.tongyi.common.validator;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 /**
  * 数据校验
