@@ -6,14 +6,14 @@ import com.tongyi.common.utils.RestResponse;
 import com.tongyi.common.xss.XssHttpServletRequestWrapper;
 import com.tongyi.core.PageInfo;
 import com.tongyi.modules.gen.service.SysGeneratorService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.commons.io.IOUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Date;
 import java.util.Map;
