@@ -22,13 +22,14 @@ import com.tongyi.modules.job.entity.ScheduleJobLogEntity;
 import com.tongyi.modules.job.service.ScheduleJobService;
 import com.tongyi.modules.job.service.ScheduleStatus;
 import com.tongyi.modules.job.utils.ScheduleUtils;
+import jakarta.annotation.PostConstruct;
 import org.quartz.CronTrigger;
 import org.quartz.Scheduler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.PostConstruct;
+
 import java.io.Serializable;
 import java.util.*;
 
