@@ -21,7 +21,7 @@ import com.tongyi.modules.sys.entity.SysMenuEntity;
 import com.tongyi.modules.sys.entity.SysUserEntity;
 import com.tongyi.modules.sys.entity.SysUserTokenEntity;
 import com.tongyi.modules.sys.service.ShiroService;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
