@@ -22,8 +22,8 @@ import com.tongyi.modules.sys.entity.SysUserEntity;
 import com.tongyi.modules.sys.service.SysRoleService;
 import com.tongyi.modules.sys.service.SysUserRoleService;
 import com.tongyi.modules.sys.service.SysUserService;
-import org.apache.commons.lang.RandomStringUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.shiro.crypto.hash.Sha256Hash;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
