@@ -12,7 +12,7 @@
 package com.tongyi.modules.oss.cloud;
 
 import com.tongyi.common.utils.DateUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.io.InputStream;
 import java.util.Date;
